start converting paginator to ts
